backend/requirements.txt
tornado==6.4
mysql-connector-python==9.1.0